[NFC] Factor duplicated type sort into separate header
Both TypeMerging.cpp and type-updating.cpp implemented topological sorts on heap
types that would visit supertypes first. Factor those implementations out into a
shared wasm-type-ordering.h header.

In principle the subtype-first sort used in subtypes.h could also be moved to
this header, but that is not yet duplicated in the code base and is more
efficient because it depends on implementation details of its current context,
so do not move it for now.
